fluent_bundle
Module bundle

FluentBundle is a collection of localization messages in Fluent.

It stores a list of messages in a single locale which can reference one another, use the same internationalization formatters, functions, scopeironmental variables and are expected to be used together.

Structs

FluentBundle

A collection of localization messages for a single locale, which are meant to be used together in a single view, widget or any other UI abstraction.